chore: remove useDidUpdate from carousel
