--- packages/blade/src/components/Collapsible/__tests__/useDidUpdate.test.ts
@@ -1,21 +1,0 @@
-import { renderHook } from '@testing-library/react-hooks';
-import { useDidUpdate } from '../useDidUpdate';
-
-describe('useDidUpdate', () => {
-  it('should invoke effectCallback after first render', () => {
-    const effectCallback = jest.fn();
-
-    const { rerender } = renderHook(
-      ({ dependencies }) => useDidUpdate(effectCallback, dependencies),
-      { initialProps: { dependencies: ['a'] } },
-    );
-
-    // Should not be called on initial render
-    expect(effectCallback).not.toHaveBeenCalled();
-
-    rerender({ dependencies: ['b'] });
-
-    // Should be called after first render, anytime dependencies change
-    expect(effectCallback).toHaveBeenCalledTimes(1);
-  });
-});
