+++ packages/blade/src/utils/useDidUpdate.native.test.ts
@@ -2,34 +2,34 @@ import { renderHook } from '@testing-library/react-native';
 import { useDidUpdate } from './useDidUpdate';
 
 describe('useDidUpdate', () => {
-  it('should run the effect only when the component is updated', () => {
-    const effect = jest.fn();
-    const { rerender } = renderHook(() => {
-      return useDidUpdate(effect, [{}]);
-    });
+  it('should invoke effectCallback after first render', () => {
+    const effectCallback = jest.fn();
 
-    expect(effect).not.toHaveBeenCalled();
+    const { rerender } = renderHook(
+      ({ dependencies }) => useDidUpdate(effectCallback, dependencies),
+      { initialProps: { dependencies: ['a'] } },
+    );
 
-    rerender({});
+    // Should not be called on initial render
+    expect(effectCallback).not.toHaveBeenCalled();
 
-    expect(effect).toHaveBeenCalledTimes(1);
-  });
+    rerender({ dependencies: ['b'] });
 
-  it('should not run the effect on the first render', () => {
-    const effect = jest.fn();
-    renderHook(() => useDidUpdate(effect, [{}]));
-
-    expect(effect).not.toHaveBeenCalled();
+    // Should be called after first render, anytime dependencies change
+    expect(effectCallback).toHaveBeenCalledTimes(1);
   });
 
   it('should not run the effect when the dependencies do not change', () => {
-    const effect = jest.fn();
-    const { rerender } = renderHook(() => useDidUpdate(effect, [1, 2, 3]));
+    const effectCallback = jest.fn();
+    const { rerender } = renderHook(
+      ({ dependencies }) => useDidUpdate(effectCallback, dependencies),
+      { initialProps: { dependencies: ['a'] } },
+    );
 
-    expect(effect).not.toHaveBeenCalled();
+    expect(effectCallback).not.toHaveBeenCalled();
 
-    rerender({});
+    rerender({ dependencies: ['a'] });
 
-    expect(effect).not.toHaveBeenCalled();
+    expect(effectCallback).not.toHaveBeenCalled();
   });
 });
